move pubsub_server
