+++ pubsub/pubsub_server.go
@@ -1,7 +1,7 @@
 // (c) 2019-2020, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package json
+package pubsub
 
 import (
 	"errors"
